Master branch now depends on patternplate-client and patternplate-server master
To release a fixed version with fixed dependencies, do the following

* create a tag in patternplate-client
* create a tag in patternplate-server
* branch patternplate from master with a single commit, that references
  the pattternplate-client and -server tags in package.json
* tag that new commit

This makes `master` an unstable branch to track as a dependency, as
intended. Use fixed tags, stream branches, or commit hashes to track a
stable dependency if you don't want your project to break through
updates. You should do this anyway.